+++ atlasdb-config/src/main/java/com/palantir/atlasdb/factory/TransactionManagers.java
@@ -333,6 +333,7 @@ public final class TransactionManagers {
                 persistentLockManager,
                 specificTableSweeper);
 
+        transactionManager.registerClosingCallback(backgroundSweeper::shutdown);
         backgroundSweeper.runInBackground();
     }