+++ atlasdb-config/src/test/java/com/palantir/atlasdb/factory/TransactionManagersTest.java
@@ -65,6 +65,7 @@ import com.palantir.atlasdb.config.ImmutableTimestampClientConfig;
 import com.palantir.atlasdb.config.ServerListConfig;
 import com.palantir.atlasdb.config.TimeLockClientConfig;
 import com.palantir.atlasdb.memory.InMemoryAtlasDbConfig;
+import com.palantir.atlasdb.transaction.impl.SerializableTransactionManager;
 import com.palantir.leader.PingableLeader;
 import com.palantir.lock.LockMode;
 import com.palantir.lock.LockRequest;
@@ -304,6 +305,21 @@ public class TransactionManagersTest {
         createLockAndTimestampServicesForConfig(config, runtimeConfig).time().getFreshTimestamp();
 
         availableServer.verify(postRequestedFor(urlEqualTo(TIMELOCK_TIMESTAMP_PATH)));
+
+    @Test
+    public void runsClosingCallbackOnShutdown() throws Exception {
+        AtlasDbConfig realConfig = ImmutableAtlasDbConfig.builder()
+                .keyValueService(new InMemoryAtlasDbConfig())
+                .defaultLockTimeoutSeconds(120)
+                .build();
+
+        Runnable callback = mock(Runnable.class);
+
+        SerializableTransactionManager manager = TransactionManagers.create(
+                realConfig, Optional::empty, ImmutableSet.of(), environment, false);
+        manager.registerClosingCallback(callback);
+        manager.close();
+        verify(callback, times(1)).run();
     }
 
     private void verifyUserAgentOnRawTimestampAndLockRequests() {